requirements.txt
# requirements.txt
pandas
httpx
authlib
python-dotenv
python-Levenshtein
rapidfuzz